Refactor into single UserProfileManager
