+++ Source/BluemixAppID/api/UserProfileManager.swift
@@ -1,4 +1,4 @@
-/* *     Copyright 2016, 2017 IBM Corp.
+/* *     Copyright 2016, 2017, 2018 IBM Corp.
  *     Licensed under the Apache License, Version 2.0 (the "License");
  *     you may not use this file except in compliance with the License.
  *     You may obtain a copy of the License at
@@ -12,7 +12,7 @@
 
 import Foundation
 
-public protocol UserAttributeManager {
+public protocol UserProfileManager {
 
     func setAttribute(key: String, value: String, completionHandler: @escaping(Error?, [String:Any]?) -> Void)
     func setAttribute(key: String, value: String, accessTokenString: String, completionHandler: @escaping(Error?, [String:Any]?) -> Void)
@@ -22,5 +22,7 @@ public protocol UserAttributeManager {
     func getAttributes(accessTokenString: String, completionHandler: @escaping(Error?, [String:Any]?) -> Void)
     func deleteAttribute(key: String, completionHandler: @escaping(Error?, [String:Any]?) -> Void)
     func deleteAttribute(key: String, accessTokenString: String, completionHandler: @escaping(Error?, [String:Any]?) -> Void)
-    
+    func getUserInfo(completionHandler: @escaping (Error?, [String: Any]?) -> Void)
+    func getUserInfo(accessTokenString: String, identityTokenString: String?, completionHandler: @escaping (Error?, [String: Any]?) -> Void)
+
 }
